Break up Alexa per functionality (#9400)
* Break up Alexa per functionality

* Lint

* Lint
