+++ tests/components/alexa/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Alexa integration."""
